# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ProbeOps 诊断 Agent API",
    "version": "0.1.0",
    "description": "T1 接口契约。P1 已实现本机 FakeLLM 基础链路，仅 fixed 策略可运行；其他策略创建返回 503 STRATEGY_UNAVAILABLE。模拟结果不构成真实根因判断。Agent 无写入或注入故障工具。Apifox project=8800905，main，Git Spec 模式，以本文件为契约源。"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "本地开发后端（P1 FakeLLM）"
    }
  ],
...
  "paths": {
...
    "/api/v1/incidents": {
      "get": {
        "operationId": "listIncidents",
        "tags": [
          "故障任务"
        ],
        "summary": "列出公开演示任务",
        "description": "规划契约；后端尚未实现。 仅访问脱敏观测，不暴露故障注入真值。",
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentList"
                }
              }
            }
          },
          "500": {
            "description": "未预期内部错误；不回显堆栈",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          }
        },
        "x-apifox-status": "developing"
      }
    },
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "operationId": "getIncident",
        "tags": [
          "故障任务"
        ],
        "summary": "读取脱敏告警与时间窗口",
        "description": "规划契约；后端尚未实现。 仅访问脱敏观测，不暴露故障注入真值。",
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          },
          "422": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          },
          "500": {
            "description": "未预期内部错误；不回显堆栈",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,64}$"
            }
          }
        ],
        "x-apifox-status": "developing"
      }
    },
    "/api/v1/strategies": {
      "get": {
        "operationId": "listStrategies",
        "tags": [
          "诊断运行"
        ],
        "summary": "列出六种诊断策略",
        "description": "规划契约；后端尚未实现。",
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyList"
                }
              }
            }
          },
          "500": {
            "description": "未预期内部错误；不回显堆栈",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          }
        },
        "x-apifox-status": "developing"
      }
    },
    "/api/v1/runs": {
      "post": {
        "operationId": "createRun",
        "tags": [
          "诊断运行"
        ],
        "summary": "创建受预算约束的诊断运行",
        "description": "规划契约；后端尚未实现。",
        "responses": {
          "202": {
            "description": "成功",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          },
          "429": {
            "description": "准入预算或并发额度不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          },
          "409": {
            "description": "状态冲突或幂等键正文不同",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          },
          "422": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          },
          "503": {
            "description": "必要依赖未就绪",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          },
          "500": {
            "description": "未预期内部错误；不回显堆栈",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 64
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRun"
              },
              "example": {
                "incident_id": "demo_001",
                "strategy_id": "competitive_cost",
                "limits": {
                  "max_steps": 12,
                  "max_llm_calls": 16,
                  "max_wall_seconds": 180,
                  "max_cost_micro_cny": 250000
                }
              }
            }
          }
        },
        "x-apifox-status": "developing"
      },
      "get": {
        "operationId": "listRuns",
        "tags": [
          "诊断运行"
        ],
        "summary": "运行列表",
        "description": "规划契约；后端尚未实现。",
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            }
          },
          "422": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          },
          "500": {
            "description": "未预期内部错误；不回显堆栈",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "x-apifox-status": "developing"
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "operationId": "getRun",
        "tags": [
          "诊断运行"
        ],
        "summary": "状态与竞争假设快照",
        "description": "规划契约；后端尚未实现。",
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          },
          "422": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          },
          "500": {
            "description": "未预期内部错误；不回显堆栈",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,64}$"
            }
          }
        ],
        "x-apifox-status": "developing"
      }
    },
    "/api/v1/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancelRun",
        "tags": [
          "诊断运行"
        ],
        "summary": "请求取消运行",
        "description": "规划契约；后端尚未实现。",
        "responses": {
          "202": {
            "description": "成功",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          },
          "422": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          },
          "500": {
            "description": "未预期内部错误；不回显堆栈",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,64}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRun"
              },
              "example": {
                "reason": "用户取消演示"
              }
            }
          }
        },
        "x-apifox-status": "developing"
      }
    },
    "/api/v1/runs/{run_id}/events": {
      "get": {
        "operationId": "listEvents",
        "tags": [
          "证据与报告"
        ],
        "summary": "增量读取事件",
        "description": "规划契约；后端尚未实现。",
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          },
          "422": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          },
          "500": {
            "description": "未预期内部错误；不回显堆栈",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,64}$"
            }
          },
          {
            "name": "after_seq",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "x-apifox-status": "developing"
      }
    },
    "/api/v1/runs/{run_id}/evidence": {
      "get": {
        "operationId": "listEvidences",
        "tags": [
          "证据与报告"
        ],
        "summary": "增量读取证据",
        "description": "规划契约；后端尚未实现。",
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidencePage"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          },
          "422": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          },
          "500": {
            "description": "未预期内部错误；不回显堆栈",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,64}$"
            }
          },
          {
            "name": "after_seq",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "x-apifox-status": "developing"
      }
    },
...
    "/api/v1/budget": {
      "get": {
        "operationId": "getBudget",
        "tags": [
          "系统"
        ],
        "summary": "读取项目预算账本",
        "description": "规划契约；后端尚未实现。",
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Budget"
                }
              }
            }
          },
          "500": {
            "description": "未预期内部错误；不回显堆栈",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "请求关联标识"
              }
            }
          }
        },
        "x-apifox-status": "developing"
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,64}$"
          },
          "retryable": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "request_id",
          "retryable"
        ]
      },
...
      "Incident": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "incident_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,64}$"
          },
          "title": {
            "type": "string"
          },
          "alert": {
            "type": "string"
          },
          "service": {
            "type": "string"
          },
          "window_start": {
            "type": "string",
            "format": "date-time"
          },
          "window_end": {
            "type": "string",
            "format": "date-time"
          },
          "dataset_version": {
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "title",
          "alert",
          "service",
          "window_start",
          "window_end",
          "dataset_version"
        ]
      },
...
      "CreateRun": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "incident_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,64}$"
          },
          "strategy_id": {
            "type": "string",
            "enum": [
              "fixed",
              "react",
              "graph_greedy",
              "competitive_cost",
              "no_cost",
              "random_probe"
            ]
          },
          "limits": {
            "$ref": "#/components/schemas/Limits"
          }
        },
        "required": [
          "incident_id",
          "strategy_id",
          "limits"
        ]
      },
...
      "IncidentList": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Incident"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "StrategyList": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Strategy"
            }
          }
        },
        "required": [
          "items"
        ]
      },
...
    }
  }
}